+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 ### Bug fixing
 - Changed all 1.0.0 versioning to v1.1.0
 - Solved https://github.com/creativetimofficial/argon-design-system-react/issues/2
-- Changes to match [Argon Design System](https://www.creative-tim.com/product/argon-design-system?ref=adsr-changelog) by Creative Tim
+- Changes to match [Argon Design System](https://www.creative-tim.com/product/argon-design-system?ref=adsr-changelog) by Abdallah Alabed
 - -  `src/components/Footers/CardsFooter.js`
 - - `src/views/IndexSections/Datepicker.js`
 - Solved build issue
 ### Major style changes
-- Changed the whole scss/css files and folders to match those of [Argon Design System](https://www.creative-tim.com/product/argon-design-system?ref=adsr-changelog) by Creative Tim
+- Changed the whole scss/css files and folders to match those of [Argon Design System](https://www.creative-tim.com/product/argon-design-system?ref=adsr-changelog) by Abdallah Alabed
 ### Deleted components
 ### Added components
 ### Deleted dependencies
@@ -51,4 +51,4 @@ Please update the following components: DateTime
 ## [1.0.0] 2019-05-13
 ### Original Release
 - Added Reactstrap as base framework
-- Added design from Argon Design System by Creative Tim
+- Added design from Argon Design System by Abdallah Alabed